requirements.txt
streamlit>=1.25.0
openai>=1.0.0
Pillow